+++ libs/base_core/src/num/float/wrapper/basic.rs
@@ -1,11 +1,13 @@
-// devela::num::float::wrapper::shared
+// devela_base_core::num::float::wrapper::basic
 //
-//! Defines all the shared, cross-platform public methods for `Float`.
+//! Defines all the basic public methods for `Float`.
 //
 // WAIT: [core_float_math](https://github.com/rust-lang/rust/issues/137578)
+// NOTE: this file is symlinked from devela_base_std::num::float::wrapper
 
+use super::definition::Float;
 use crate::{
-    _FloatInternals, Cmp, Float, FloatCategory, Sign, concat as cc, is, stringify as sfy, whilst,
+    _FloatInternals, Cmp, FloatCategory, Sign, concat as cc, is, stringify as sfy, whilst,
 };
 
 /// Implements methods independently of any features
@@ -14,18 +16,18 @@ use crate::{
 /// $uf:  unsigned int type with the same bit-size.
 /// $ie:  signed int type used for integer exponentiation.
 /// $ue:  unsigned int type used for integer exponentiation and number of terms (u32).
-macro_rules! impl_float_shared {
+macro_rules! impl_float_basic {
     () => {
-        impl_float_shared![ (f32:u32, i32, u32), (f64:u64, i32, u32)];
+        impl_float_basic![ (f32:u32, i32, u32), (f64:u64, i32, u32)];
         // #[cfg(feature = "nightly_float")] // TODO
-        // impl_float_shared![ (f16:u16, u32), (f128:u128, u32)];
+        // impl_float_basic![ (f16:u16, u32), (f128:u128, u32)];
     };
 
     ($( ($f:ty:$uf:ty, $ie:ty, $ue:ty)),+) => {
-        $( impl_float_shared![@$f:$uf, $ie, $ue]; )+
+        $( impl_float_basic![@$f:$uf, $ie, $ue]; )+
     };
     (@$f:ty:$uf:ty, $ie:ty, $ue:ty) => {
-        /// # *Common implementations with or without `std`*.
+        /// # *Common implementations with or without `std`.*
         impl Float<$f> {
             /// The largest integer less than or equal to itself.
             /// # Formulation
@@ -221,7 +223,7 @@ macro_rules! impl_float_shared {
             #[doc = crate::_FLOAT_FORMULA_SCALE!()]
             /// # Examples
             /// ```
-            /// # use devela::Float;
+            /// # use devela_base_core::Float;
             #[doc = cc!["assert_eq![Float(45_", sfy![$f], ").scale(0., 360., 0., 1.), 0.125];"]]
             #[doc = cc!["assert_eq![Float(45_", sfy![$f], ").scale(0., 360., -1., 1.), -0.75];"]]
             #[doc = cc!["assert_eq![Float(0.125_", sfy![$f], ").scale(0., 1., 0., 360.), 45.];"]]
@@ -241,7 +243,7 @@ macro_rules! impl_float_shared {
             #[doc = crate::_FLOAT_FORMULA_LERP!()]
             /// # Example
             /// ```
-            /// # use devela::Float;
+            /// # use devela_base_core::Float;
             #[doc = cc!["assert_eq![Float(0.5_", sfy![$f], ").lerp(40., 80.), 60.];"]]
             // TODO more examples extrapolated
             /// ```
@@ -376,7 +378,7 @@ macro_rules! impl_float_shared {
             ///
             /// # Example
             /// ```
-            /// # use devela::Float;
+            /// # use devela_base_core::Float;
             #[doc = cc!["assert_eq![Float(50.0_", sfy![$f], ").clamp(40., 80.), 50.];"]]
             #[doc = cc!["assert_eq![Float(100.0_", sfy![$f], ").clamp(40., 80.), 80.];"]]
             #[doc = cc!["assert_eq![Float(10.0_", sfy![$f], ").clamp(40., 80.), 40.];"]]
@@ -403,7 +405,7 @@ macro_rules! impl_float_shared {
             ///
             /// # Example
             /// ```
-            /// # use devela::Float;
+            /// # use devela_base_core::Float;
             #[doc = cc!["assert_eq![Float(50.0_", sfy![$f], ").clamp_total(40., 80.), 50.];"]]
             #[doc = cc!["assert_eq![Float(100.0_", sfy![$f], ").clamp_total(40., 80.), 80.];"]]
             #[doc = cc!["assert_eq![Float(10.0_", sfy![$f], ").clamp_total(40., 80.), 40.];"]]
@@ -431,7 +433,7 @@ macro_rules! impl_float_shared {
             ///
             /// # Example
             /// ```
-            /// # use devela::Float;
+            /// # use devela_base_core::Float;
             #[doc = cc!["assert_eq![Float(50.0_", sfy![$f], ").clamp_nan(40., 80.), 50.];"]]
             #[doc = cc!["assert_eq![Float(100.0_", sfy![$f], ").clamp_nan(40., 80.), 80.];"]]
             #[doc = cc!["assert_eq![Float(10.0_", sfy![$f], ").clamp_nan(40., 80.), 40.];"]]
@@ -445,7 +447,7 @@ macro_rules! impl_float_shared {
             ///
             /// # Example
             /// ```
-            /// # use devela::Float;
+            /// # use devela_base_core::Float;
             #[doc = cc!["assert_eq![Float(50.0_", sfy![$f], ").max_nan(80.), 80.];"]]
             #[doc = cc!["assert_eq![Float(100.0_", sfy![$f], ").max_nan(80.), 100.];"]]
             /// ```
@@ -469,7 +471,7 @@ macro_rules! impl_float_shared {
             ///
             /// # Example
             /// ```
-            /// # use devela::Float;
+            /// # use devela_base_core::Float;
             #[doc = cc!["assert_eq![Float(50.0_", sfy![$f], ").min_nan(80.), 50.];"]]
             #[doc = cc!["assert_eq![Float(100.0_", sfy![$f], ").min_nan(80.), 80.];"]]
             /// ```
@@ -522,7 +524,7 @@ macro_rules! impl_float_shared {
             ///
             /// # Examples
             /// ```
-            /// # use devela::Float;
+            /// # use devela_base_core::Float;
             /// let coefficients = [2.0, -6.0, 2.0, -1.0];
             #[doc = cc!["assert_eq![Float(3.0_", sfy![$f], ").eval_poly(&coefficients), 5.0];"]]
             #[doc = cc!["assert_eq![Float(3.0_", sfy![$f], ").eval_poly(&[]), 0.0];"]]
@@ -608,4 +610,4 @@ macro_rules! impl_float_shared {
         }
     };
 }
-impl_float_shared!();
+impl_float_basic!();