+++ libs/base_core/src/text/fmt/buf.rs
@@ -46,7 +46,7 @@ macro_rules! format_buf {
 pub use format_buf;
 
 #[doc = crate::_TAG_FMT!()]
-/// A specialized formatter that works with a fixed byte buffer and has truncation detection.
+/// A specialized formatter with a fixed byte buffer and truncation detection.
 ///
 /// It should be faster at runtime than the default formatter.
 #[derive(Debug)]