+++ libs/base_std/src/lib.rs
@@ -28,6 +28,7 @@ items! {
     pub mod build;
     pub mod code;
     pub mod data;
+    pub mod num;
     pub mod phys;
     pub mod sys;
     pub mod text;
@@ -51,6 +52,7 @@ pub mod zall {
         build::_all::*,
         code::_all::*,
         data::_all::*,
+        num::_all::*,
         phys::_all::*,
         sys::_all::*,
         text::_all::*,