+++ libs/base_std/src/num/float/mod.rs
@@ -1,0 +1,16 @@
+// devela_base_std::num
+//
+#![doc = crate::_DOC_NUM!()]
+//
+// safety
+#![cfg_attr(base_safe_num, forbid(unsafe_code))]
+
+mod wrapper;
+
+crate::structural_mods! { // _mods
+    _mods {
+        pub use super::{
+            wrapper::_all::*,}
+        ;
+    }
+}