+++ libs/base_std/src/num/float/wrapper/consts.rs
@@ -1,0 +1,1 @@
+../../../../../../libs/base_core/src/num/float/wrapper/consts.rs