+++ libs/base_std/src/num/mod.rs
@@ -1,0 +1,12 @@
+// devela_base_std::num::float
+//
+//!
+//
+
+mod float;
+
+crate::structural_mods! { // _mods
+    _mods {
+        pub use super::float::*;
+    }
+}