+++ src/num/float/mod.rs
@@ -4,7 +4,6 @@
 //
 
 mod ext_float; // FloatExt
-mod wrapper; // Float
 
 // re-exports
 crate::mod_path!(_c "../../../libs/base_core/src/num/float/reexports.rs");
@@ -13,15 +12,22 @@ crate::structural_mods! { // _mods
     _mods {
         pub use super::{
             ext_float::*,
-            wrapper::*,
         };
 
         // re-exports
         pub use super::_c::*;
         #[doc(inline)]
-        pub use devela_base_core::{
+        pub use devela_base_core::num::{
             FloatConst,
             f32bits, f32bits_niche, f64bits, f64bits_niche, fsize,
         };
+
+        #[cfg(not(feature = "std"))]
+        #[cfg_attr(nightly_doc, doc(auto_cfg(hide(feature = "std"))))]
+        pub use devela_base_core::num::Float;
+
+        #[cfg(feature = "std")]
+        #[cfg_attr(nightly_doc, doc(auto_cfg(hide(feature = "std"))))]
+        pub use devela_base_std::num::FloatStd as Float;
     }
 }